requirements.txt
streamlit
pandas
matplotlib
seaborn
arabic-reshaper
python-bidi

